src/test/java/com/example/demo/domain/contract/service/test/ContractServiceTest.java
package com.example.demo.domain.contract.service.test;

import com.example.demo.TechnicalError;
import com.example.demo.domain.contract.model.Contract;
import com.example.demo.domain.contract.model.Term;
import com.example.demo.domain.contract.service.ContractService;
import com.example.demo.infrastructure.model.repository.ContractRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import java.math.BigDecimal;
import java.util.Arrays;
import java.util.List;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.*;

public class ContractServiceTest {

    @Mock
    private ContractRepository contractRepository;

    @InjectMocks
    private ContractService contractService;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void testCreateContract() {
        // Given
        BigDecimal amount = BigDecimal.valueOf(1000);
        List<com.example.demo.aggregate.Contract> aggregateContracts = Arrays.asList(/* mock aggregate contracts */);
        when(contractRepository.saveAll(any())).thenReturn(aggregateContracts);

        // When
        List<Contract> contracts = contractService.createContract(amount);

        // Then
        assertNotNull(contracts);
        // Add assertions to verify the behavior of createContract method
    }

    @Test
    public void testCalculateContract() {
        // Given
        BigDecimal amount = BigDecimal.valueOf(1000);

        // When
        List<Contract> contracts = contractService.calculateContract(amount);

        // Then
        assertNotNull(contracts);
        // Add assertions to verify the behavior of calculateContract method
    }

    @Test
    public void testDeleteContract() {
        // Given
        Long id = 1L;
        doNothing().when(contractRepository).deleteById(anyInt());

        // When
        assertDoesNotThrow(() -> contractService.deleteContract(id));

        // Then
        // Add assertions to verify the behavior of deleteContract method
    }

    @Test
    public void testDeleteContractWithNegativeId() {
        // Given
        Long id = -1L;

        // When
        TechnicalError technicalError = assertThrows(TechnicalError.class, () -> contractService.deleteContract(id));

        // Then
        assertEquals("teh value provided is negative", technicalError.getMessage());
    }

    @Test
    public void testUpdateContractAmount() {
        // Given
        Long id = 1L;
        BigDecimal amount = BigDecimal.valueOf(1500);
        com.example.demo.aggregate.Contract contract = new com.example.demo.aggregate.Contract(/* initialize contract */);
        when(contractRepository.existsById(anyInt())).thenReturn(true);
        when(contractRepository.findById(anyInt())).thenReturn(java.util.Optional.of(contract));
        when(contractRepository.save(any())).thenReturn(contract);

        // When
        assertDoesNotThrow(() -> contractService.updateContractAmount(id, amount));

        // Then
        // Add assertions to verify the behavior of updateContractAmount method
    }

    // Add more test methods for other scenarios as needed
}

